scripts/requirements.txt
discord.py
pygame
PyOpenGL
pywavefront
selenium
keyboard